+++ tests/laravel/app/Console/Kernel.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Console;
+
+use Butler\Service\Console\Kernel as ConsoleKernel;
+use Illuminate\Console\Scheduling\Schedule;
+
+class Kernel extends ConsoleKernel
+{
+    protected function schedule(Schedule $schedule): void
+    {
+        //
+    }
+
+    protected function commands(): void
+    {
+        $this->load(__DIR__ . '/Commands');
+    }
+}