+++ tests/laravel/routes/api.php
@@ -1,0 +1,5 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+
+Route::get('/dummy-api-route', fn () => true)->name('dummy-api-route');