+++ tests/laravel/routes/web.php
@@ -1,0 +1,5 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+
+Route::get('/dummy-web-route', fn () => true)->name('dummy-web-route');